+++ lib/rubocop.rb
@@ -27,6 +27,7 @@ require 'rubocop/cop/check_methods'
 
 require 'rubocop/cop/lint/assignment_in_condition'
 require 'rubocop/cop/lint/block_alignment'
+require 'rubocop/cop/lint/condition_position'
 require 'rubocop/cop/lint/debugger'
 require 'rubocop/cop/lint/empty_ensure'
 require 'rubocop/cop/lint/end_alignment'